requirements.txt
groq>=0.9.0
fastapi
uvicorn[standard]
httpx
python-dotenv
